// src/app/models/master-dealer-manager-form.model.ts
export interface MasterDealerManagerFormModel {
  dealerId: number;
  name: string;
  birthDate: string;
  cpf: string;
  rg: string;
  genre: string;
  email: string;
  phone: string;
  storeId: string;
  postalCode: string;
  address: string;
  addressNumber: string;
  neighborhood: string;
  city: string;
  state: string;
  complement: string;

}